feat: added background template in generated pdf file. fix: some refactoring has been performed.
